refactor: Split cli.py into cli/ package (#131)
- cli/__init__.py: Entry point (main, exception handling)
- cli/parser.py: Argument parser (create_parser)
- cli/commands.py: Command implementations (cmd_*)
- cli/formatters.py: Output formatting (format_forecast_*, WEATHER_EMOJI)
- cli/helpers.py: Shared helpers (get_forecast_source, get_historical_source)
- cli/__main__.py: Support for python -m pvforecast.cli

Improves maintainability through separation of concerns.
All tests passing.
