+++ src/pvforecast/cli/__main__.py
@@ -1,0 +1,8 @@
+"""Entry point for `python -m pvforecast.cli`."""
+
+import sys
+
+from . import main
+
+if __name__ == "__main__":
+    sys.exit(main())
